Store away the old script command and start migrating to the new data structure
